requirements.txt
streamlit>=1.55
pandas>=2.0
plotly>=5.20
yfinance>=0.2
Pillow>=10.0
rapidocr-onnxruntime>=1.3
